arm64: Add a fixup handler for alignment faults in aarch64 code
A later patch will hand out Device memory in some cases to code
which expects a Normal memory type, as an errata workaround. Unaligned
accesses to Device memory will fault though, so here we add a fixup handler
to emulate faulting accesses, at a performance penalty.

Many of the instructions in the Loads and Stores group are supported,
but these groups are not handled here:

 * Advanced SIMD load/store multiple structures
 * Advanced SIMD load/store multiple structures (post-indexed)
 * Advanced SIMD load/store single structure
 * Advanced SIMD load/store single structure (post-indexed)
 * Load/store memory tags
 * Load/store exclusive
 * LDAPR/STLR (unscaled immediate)
 * Load register (literal) [cannot Alignment fault]
 * Load/store register (unprivileged)
 * Atomic memory operations
 * Load/store register (pac)

Instruction implementations are translated from the Exploration tools'
ASL specifications.

Signed-off-by: D Scott Phillips <scott@os.amperecomputing.com>
Signed-off-by: Huang Shijie <shijie@os.amperecomputing.com>
Signed-off-by: John Gong <john.gong@amperecomputing.com>