tcp: fix issues when enabling tcp_wan_timestamps feature.
To some extent, It can solve/avoid sending a syn packet including
timestamp which could cause the wscale to be encoded into the
timestamp on the other side. Most importantly, it handles the
missing side case (using this feature as a client).

However, it still leave one minor/race issue unsolved which I don't
have intention to get done for now: If the client has done sending
a syn packet, user sets tcp_wan_timestamps to 0 to enable this
feature, it will cause the same issue as before. However, chance is
more slim and we can detect it easily.

Fixes: 682b7045e5204 ("net: add net.ipv4.tcp_wan_timestamps sysctl to switch timestamps function")
Signed-off-by: Jason Xing <kernelxing@tencent.com>