crypto: qat - move PFVF compat checker to a function
commit 867e801005e9 ("crypto: qat - move PFVF compat checker to a function") upstream

Move the code that implements VF version compatibility on the PF side to
a separate function so that it can be reused when doing VM live
migration.

This does not introduce any functional change.

Intel-SIG: commit 867e801005e9 ("crypto: qat - move PFVF compat checker to a function)"
Backport to support Intel QAT live migration for in-tree driver

Signed-off-by: Xin Zeng <xin.zeng@intel.com>
Reviewed-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
[ Aichun Shi: amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>