crypto: qat - add interface for live migration
commit 0fce55e5334d ("crypto: qat - add interface for live migration") upstream

Extend the driver with a new interface to be used for VF live migration.
This allows to create and destroy a qat_mig_dev object that contains
a set of methods to allow to save and restore the state of QAT VF.
This interface will be used by the qat-vfio-pci module.

Intel-SIG: commit 0fce55e5334d ("crypto: qat - add interface for live migration)"
Backport to support Intel QAT live migration for in-tree driver

Signed-off-by: Xin Zeng <xin.zeng@intel.com>
Reviewed-by: Giovanni Cabiddu <giovanni.cabiddu@intel.com>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
[ Aichun Shi: amend commit log ]
Signed-off-by: Aichun Shi <aichun.shi@intel.com>