net: hns3: fix port vlan filter not disabled problem in dynamic vlan mode
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I9D6UN
CVE: NA

----------------------------------------------------------------------

The driver set port vlan filter to "on" when initialized, and
then set port vlan filter to "off" in service task if in dynamic
vlan mode. Between this two settings, if user set a vlan to the
port, the port vlan filter is remain in "on" status, which is
not expected.

To fix the problem, if in dynamic vlan mode, set the port vlan
filter to "off".

Fixes: 184cd221a863 ("net: hns3: disable port VLAN filter when support function level VLAN filter control")
Fixes: 2ba306627f59 ("net: hns3: add support for modify VLAN filter state")
Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Signed-off-by: Hao Chen <chenhao418@huawei.com>
Signed-off-by: hongrongxuan <hongrongxuan@huawei.com>