emm: mm, block: restore rw_page infrastructure with failure tolerance
Upstream: no

Our ZRAM module will still use rw_page and may return error for certain
writeback, so this interface is needed. It was removed in upstream
commit 3222d8c2a7f8 ("block: remove ->rw_page"), add it back.

Also this is a bit different from upstream version, if rw_page failed
with return value EOPNOTSUPP, fall back to bio. Else consider the whole
IO failed, to work better with ZRAM. We'll rename this to rw_folio later
to make it more distinguishable.

Signed-off-by: Kairui Song <kasong@tencent.com>