dist: tk: enable keypkg by default
For non-tagged build, keypkg contains the build time key. Non-tagged
build are not suppose to be used for production servers so this only
making debugging easier.

For tagged koji build, keypkg contains a dummy empty key, sign-file is
hooked by tencent-kernel-signer so it's safe to do so too.

If for some special purpose the keypkg shouldn't be exposed to users,
just don't add the kernel-signing-keys to repo.

Signed-off-by: Kairui Song <kasong@tencent.com>